fixed styling, refactored uploader
