+++ app/frontend/components/shared/chefs/styles.scss
@@ -1,8 +1,8 @@
-@import "formiojs/dist/formio.form.css";
 @import url("https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900");
 @import url("https://fonts.googleapis.com/css?family=Material+Icons");
 @import url("https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css");
 @import "bootstrap/dist/css/bootstrap.min.css";
+@import "formiojs/dist/formio.form.css";
 
 // Variables
 $bcgov-blue: #003366;
@@ -359,9 +359,5 @@ h2 {
 
   .card {
     margin-bottom: 24px !important;
-  }
-
-  .fa::before {
-    display: none;
   }
 }
